Small refactor for better readability. Separated spec from string_calculator in two different files
